OneClickSelect, Most orders implemented including transporting resources
Patrol is P, Stop is S, Guard is G (not completely done) Follow is F
There are resources in the starting factory which your camera is
positioned facing towards to begin with
Click on an item in one unit's inventory then click on another unit's
inventory while both are selected to issue a transfer order. One of the
units must be a transporter (worker). The leftmost button in the
inventory is a summary of the inventory quantity/max (max not
implemented). You can click it to select the unit.