SafeZone Generation, Main Menu + More!
World Generation is continued by the addition of random start locations
and safe zone generation. World size, number of players, ratio of safe
zone to world size, and random seed are all fully functional modifiers
for world generation.
SafeZone is currently just some raised terrain, but plans are to have it
be a ring of cliffs instead of a filled plateau.
Added Main Menu with Start Button. This is the scene you need to run for
the game to work.
Added loading screen to make waiting for world generation a little more
interesting.
Disabled day/night cycle because it was annoying (too dark at night,
need to fix)
Chunk Size lowered from 512 to 128 to help with faster world generation
for testing purposes.