More Players, Bug Fixes
Commanders now spawn for all players start locations, with up to 8
colors.
Fixed the placement of safe zones to properly corespond to start
locations (once I placed the commanders it was easy to see I was
improperly inverting the indecies of the safe zone height bitmap check)
Finally fixed the gap issue in setting multi-terrain heights... turns
out height maps always have a resolution of a power of 2 + 1 (even if
you try to set them to something else), so we need to ensure the chunk
size is a power of 2 + 2 (since our resolution is ChunkSize/2)
Fixed the loading screen background since the planet image was centered
instead of stretch.