Start Location Improvements, Bug Fix
Doing some real math now to optimize the initial start location guesses,
improving the odds of generating a world quickly (minDistance used to be
basically 0, now it's half of the radius of a circle which would have
startSafeZoneArea).
Workers no longer need to return to factory to get resources when
ordered to build something when they have exactly what is needed.