Unity Update, Command Grid, Nested Prefabs
Updated Unity to 20183.04b for nested prefabs so we could split out the
unit model gameobject from the unit's script gameobject. This lets us
instantiate a copy of the unit model on the UI for display.

Mostly finished the command grid (still need back buttons), allows for
swapping out command lists for different objects. Uses the active
subgroup to decide which abilities to display. WASD is still camera
movement as edge-scrolling in a non-fullscreen (ie unity editor) is a
pain. Later I'll add edge scrolling and probably move WASD to HJKL or
something.

Refactored the box selection code for readability and a bit of
performance.
Fixed onSelectionChange event invoking multiple times for a
single select.
In the future, may want to consider a BSP tree for selection instead of
iterating over all objects and checking position. This may help when object counts go into the many thousands per active simulation.