Updated Unity, Added Status Bars
Previous version of unity was unstable, moved to unity 2018.3.0b5.

When selecting a single unit you can now see status bars for the unit's
defenses and energy systems in above/below the model in the UI bar.
The status bars keep to the order in "InOrderDefenses" (order used for
distributing damage to units) and the new
"InOrderEnergySystemsWithStorage" (a little verbose, potentially used
for order of consumption of enegry in the future).

Renamed Hull to Health, HullWithArmor to Armor, added currentHitPoints
and maxHitPoints to Defense.

Disabled the construction menu at the bottom of the screen since it is
no longer needed. Will remove ButtonManager and MenuManager when the
UIBar has a way to display inventories.